chore: removed gmail agent and clean up remaining gmail mcp server code, refactored templates
