--- src/mastra/mcpservers/gmail.ts
@@ -1,11 +1,0 @@
-import { MCPClient } from "@mastra/mcp";
-
-// Configure MCPClient to connect to your server(s)
-export const gmailMcp = new MCPClient({
-  servers: {
-    gmail: {
-      command: "npx",
-      args: ["@gongrzhe/server-gmail-autoauth-mcp"],
-    },
-  },
-});
